+++ lib/screens/me/widgets/propose_card_deliverable.dart
@@ -81,7 +81,7 @@ class ProposeCardDeliverable extends ConsumerWidget {
           PrimaryProposeActions(id: id, onResolved: onResolved)
         else
           StalledProposeActions(
-            id: id,
+            attention: attention,
             onResolved: onResolved,
             viewSourceLabel: 'View deliverable',
             onViewSource: deliverableId.isEmpty